Enhanced the status event analysis to do the same for all label categories
