+++ run.py
@@ -13,7 +13,7 @@ from analysis.event_analysis import EventAnalysis
 from analysis.issue_analysis import IssueAnalysis
 from analysis.user_specific_issue_analysis import UserSpecificIssueAnalysis
 from analysis.label_trend_analysis import LabelTrendAnalysis
-from analysis.status_event_analysis import StatusEventAnalysis
+from analysis.event_label_categories_analysis import EventLabelCategoriesAnalysis
 
 def parse_args():
     """
@@ -62,7 +62,7 @@ elif args.feature == 4:
 elif args.feature == 5:
     LabelTrendAnalysis().run()
 elif args.feature == 6:
-    StatusEventAnalysis().run()
+    EventLabelCategoriesAnalysis().run()
     
 else:
     print('Need to specify which feature to run with --feature flag.')
